[Fix]: Corrected cron job of certbot renewal automation
Ensure that certbot containers destroy on exit to prevent perpetual orphan generation. Simultaneously remove any existing orphaned containers as a precautionary action. Make sure that the nginx container has a folder for the acme challenge by creating it in the docker file. Link the acme endpoint to the shared mount/volume between the nginx and certbot containers.